requirements.txt
matplotlib==3.7.0
mtcnn==0.1.1
numpy==1.24.2
opencv_python==4.8.0.74
tensorflow_macos==2.13.0
tqdm==4.65.0
